#625 rename GraphTypeRegistries and move them to the same location
